feat: nova classe ReservtionService e o modificado o service de Room  #5

+++ src/main/java/com/luizguilherme/meeting_management/service/RoomService.java
@@ -1,12 +1,14 @@
 package com.luizguilherme.meeting_management.service;
 
 import com.luizguilherme.meeting_management.model.Room;
+import com.luizguilherme.meeting_management.repository.ReservationRepository;
 import com.luizguilherme.meeting_management.repository.RoomRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Optional;
 
@@ -15,6 +17,9 @@ public class RoomService {
 
     @Autowired
     private RoomRepository roomRepository;
+
+    @Autowired
+    private ReservationRepository reservationRepository;
 
     public List<Room> getAllRooms() {
         return roomRepository.findAll();
@@ -39,10 +44,18 @@ public class RoomService {
                     room.setCapacity(roomDetails.getCapacity());
                     room.setResources(roomDetails.getResources());
                     return roomRepository.save(room);
-                }).orElseThrow(() -> new RuntimeException("Room not found"));
+                }).orElseThrow(() -> new RuntimeException("Sala não foi encontrada"));
     }
 
     public void deleteRoom(Long id) {
-        roomRepository.deleteById(id);
+        boolean hasFutureReservations = reservationRepository
+                .findByRoomIdAndStartTimeBetween(id, LocalDateTime.now(), LocalDateTime.MAX)
+                .isEmpty();
+
+        if (hasFutureReservations) {
+            roomRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("Não é possível deletar a sala, pois há reservas ativas associadas.");
+        }
     }
 }
